requirements.txt
Django==5.2.5
djangorestframework==3.16.1
Pillow>=10.4.0
psycopg[binary]>=3.2.0
django-ckeditor>=6.7.1
django-cors-headers